Move Videos feature to vertical-slice folders
Co-locates the Videos controller, services, interfaces, DTOs, and tests
under Features/Videos/ folders in each project (API, Application,
Contracts, Tests). Adds VideosModule for per-feature DI registration
and VideoRoutes for per-feature route constants.

Converter-facing contracts (UpdateVideoInfoRequest, VideoToTransformResponse,
GetPublishSasResponse, UploadConvertedVideoResponse) remain in the original
Requests/Responses folders; they will move with the upcoming Conversion slice.
