+++ src/backend/Application/Features/Videos/VideosModule.cs
@@ -1,0 +1,16 @@
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Application.Features.Videos;
+
+public static class VideosModule
+{
+    extension(IServiceCollection services)
+    {
+        public IServiceCollection AddVideosFeature()
+        {
+            services.AddScoped<IVideoService, VideoService>();
+            services.AddScoped<IVideoUploaderService, VideoUploaderService>();
+            return services;
+        }
+    }
+}
