+++ src/backend/TB.DanceDance.API.Contracts/Features/Groups/GroupWithVideosResponse.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Collections.Generic;
-using TB.DanceDance.API.Contracts.Responses;
+using TB.DanceDance.API.Contracts.Features.Videos;
 
 namespace TB.DanceDance.API.Contracts.Features.Groups
 {
