+++ src/backend/TB.DanceDance.API.Contracts/Features/Videos/UpdateVideoCommentSettingsRequest.cs
@@ -1,6 +1,6 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace TB.DanceDance.API.Contracts.Requests
+namespace TB.DanceDance.API.Contracts.Features.Videos
 {
     public class UpdateVideoCommentSettingsRequest
     {
