+++ src/backend/TB.DanceDance.API.Contracts/Features/Videos/UploadVideoInformationResponse.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace TB.DanceDance.API.Contracts.Models
+namespace TB.DanceDance.API.Contracts.Features.Videos
 {
     public class UploadVideoInformationResponse
     {
