+++ src/backend/TB.DanceDance.API/Features/Sharing/ShareController.cs
@@ -1,4 +1,5 @@
 using Application.Features.Sharing;
+using Application.Features.Videos;
 using Domain.Services;
 using Infrastructure.Identity.IdentityResources;
 using Microsoft.AspNetCore.Authorization;
