+++ src/tests/TB.DanceDance.Mobile.Tests/IntegrationTests/VideoUploaderTests.cs
@@ -2,6 +2,7 @@
 using Microsoft.Maui.Devices;
 using NSubstitute;
 using System.Threading.Channels;
+using TB.DanceDance.API.Contracts.Features.Videos;
 using TB.DanceDance.API.Contracts.Models;
 using TB.DanceDance.API.Contracts.Requests;
 using TB.DanceDance.Mobile.Library.Data;
