+++ src/tests/TB.DanceDance.Tests/Features/Sharing/SharedLinkServiceTests.cs
@@ -1,8 +1,6 @@
 using Application.Features.AccessManagement;
 using Application.Features.Sharing;
-using Application.Services;
 using Domain.Entities;
-using Domain.Services;
 using Infrastructure.Data;
 using TB.DanceDance.Tests.TestsFixture;
